# Make several improvements to binary page allocator